+++ src/main/java/com/playtomic/tests/wallet/api/dtos/WalletDto.java
@@ -1,6 +1,7 @@
 package com.playtomic.tests.wallet.api.dtos;
 
 public record WalletDto(
+        String walletId,
         int balance,
         String currency
 ) {