+++ src/main/java/com/playtomic/tests/wallet/repositories/wallet/WalletRepository.java
@@ -1,5 +1,7 @@
 package com.playtomic.tests.wallet.repositories.wallet;
 
+import com.playtomic.tests.wallet.domain.Wallet;
+
 import java.util.Optional;
 
 public interface WalletRepository {
@@ -7,4 +9,6 @@ public interface WalletRepository {
     void addBalanceLine(String walletId, int amount, String concept);
 
     Optional<String> findWalletIdForUserId(String userId);
+
+    Optional<Wallet> getWallet(String walletId);
 }